Break up vore panel UI into separate tsx files
